pyproject.toml
[project]
name = "rmq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["snowflake-connector-python>=3.15.0", "snowflake-core>=1.6.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
